web: Don't show primary address for password reset
Only show the actual address used in the flash message if it was
an address which the user entered specifically.  If the email
address was taken from the user's profile, just show "your primary
address" instead, to avoid revealing the user's email address
to anyone who enters their user name on the password reset page.