db: Add affiliation (and inst.) to member table
Affiliation has been added to the member table to allow proposal
editors to specify the affiliation of each member of the proposal.
This is now a required field (not allowed to be NULL).
The web interface for creating a proposal now needs to ask the
creator of the proposal to specify their own affiliation.

Institution has been added to the member table to allow institution
assignments to be frozen when the proposal's call closes.  It is allowed
to be NULL (for while the call is open) and is currently just left as
NULL.