db: Include person and institution with members
The "Member" records returned by the "search_member" method now
include the person's name an information about their institution.
The actual "institution_id" field from the database is excluded,
and instead a "resolved_institution_id" is given: this is the person's
institution_id unless the row of the member table actually includes
a specific institution_id, in which case it is that.  The institution
information given should be for this resolved institution.  The
plain "institution_id" column isn't returned in order to prevent
the wrong one being used by mistake.