web: Add a page allowing members to be added
This page, like the person insitution choice page, has two forms:
one for selecting an existing person from the directory
and another for entering the details of an unlisted person.

Somewhat unsatisfactorily, common elements (affiliation, editor
and observer status) are repeated between both forms.  It does
seem necessary for there to be two separate forms, however,
so that the default submit button for the "new member" part
of the page is the correct one.  (Otherwise pressing return
in the new member name or email text field will submit
the form using the default submit button from the "from directory"
part of the form, causing whoever is selected in the directory
selection list to be added to the proposal.)