test: Randomize autoincrement start numbers
The database tests now use randomized starting numbers for the
autoincrement fields, to avoid failing to detect the usage of an
incorrect foreign key due to the same key being present in multiple
tables.  The previous bad value, 999, has been replaced with
1999999 to be outside the range of randomized values (1-1000000).

Doing this for the purpose of the test suite via the sqlite_sequence
table required turning on the sqlite_autoincrement table option.