db: Delete proposal PDF/text when alternative set
If a proposal PDF is uploaded then any existing text (and figures)
for the same role is deleted.  Likewise if new text is written then
any existing PDF is removed.

This has been done at the database interface level so that the
new data is inserted in the same transaction which removes the
old data.  It should ensure that a proposal doesn't end up
with both of the alternative types of technical or scientific
justification.