web: Allow the browser to cache proposal resources
The "send_file" decorator has a new "allow_cache" argument, which if set,
will add a Cache-Control header to the response enabling caching.

This commit takes the simplistic approach of just adding the MD5 sum
of the resource to the URL.  The uploaded figure or PDF's MD5 sum is
used for the figure view, preview and thumbnail and for the PDF previews.
The handler functions now all check that the MD5 sum is correct before
serving the resource.  The advantage is that the MD5 sum identifying
the version of each resource is written directly into the HTML being
sent to the browser, so it can determine with no additional checking
whether or not it needs to fetch the resource.  With this commit in
place, loading the proposal view simply loads the templated HTML,
whereas previously all the figures and PDF previews would be retrieved
on every load.