db: Take MOC order when adding/updating a MOC
The MOC is then normalized at this order before being stored in
the database.  This is to ensure that the caller does consider
the order of the moc -- it must not be higher that the order
at which cell searches will be performed because that would
lead to cells at the higher orders being missed.