web: Add name resolve button to target editor
Ideally this would contact CADC directly, but their service
currently does not support HTTPS or CORS, so it is necessary
to proxy the request through Hedwig.  The proxy route should
behave like the CADC service, so the JavaScript could easily be
altered to query CADC directly when this is possible.