db: Store addresses directly in message_recipient
When a message recipient has a specific email address, it is now stored
directly in the message_recipient table rather than via a reference
to the email table.  This is so that we can be sure that the email is
sent to the specified address even if the user changes the corresponding
address on their profile.