web: Avoid displaying unnecessary ITC values
Now calculations' "parsed input" (which is what is saved in the
calculation table) is None for values which are not relevant given
other selections.  (E.g. raster size in jiggle mode, or pixel size
in matched-filter mode.)  Then hide the display of values which
are None on the proposal.

Also abbreviated boolean values so that they just appear if true
and are not shown otherwise.  (I.e. the value isn't shown for
boolean and the whole parameter is hidden if it's false.)