jcmt: Set the het. ITC freq. res. HTML min to 0
This stops it from going negative (in Firefox at least) when the
up and down buttons are pressed.  Unfortunately it seems that you
can't specify an increment for those buttons without limiting the
allowed values to particular steps, so the step is left at "any"
for now.  It could perhaps be changed to "0.001" so that the buttons
behave more sensibly.