web: Display "back to proposal" on new calculations
Requested by HALP so that you can navigate back from the ITCs
to your proposal without using "Save and return to proposal".
This requires an extra parameter to be tracked so that we can
show the link only when the current proposal identifier is the
same as that originally used to show the calculator.  This
new parameter is also set when a calculation is saved -- so
that having saved a calculation to a proposal you then see
a link back to that proposal.