web: Allow changing calc. mode with missing input
Now provide the defaults (for the current mode) to the parse_input
method when attempting to change mode.  This allows the default
values to be used for the mode-changing calculation and avoids
"Invalid value" errors due to missing parameters.

The defaults currently aren't used in place of present but invalid
parameters so as not to loose a parameter which a user has partially
entered.