db: Add connection pool checks
Based on the examples in the SQLAlchemy documentation, this
commit introduces checks for stale connections and connections
shared between processes.  (The latter is in case the system is
run with multiple WSGI daemon processes.)  It also sets a pool
recycle time of one hour to try to avoid stale connections
which result from MySQL dropping inactive connections.