facility: Add methods to parse proposal codes
The parse_proposal_code method has been split into two parts: the
public part which takes a code and looks up the parsed parts of the
code in the database, and a private part which actually parses
the code.  This is so that, in the common case of a proposal
code comprising the semester and queue codes and proposal number,
only the private part needs to be implemented for each facility.