db/web: Add text format option to descriptions
Added a text format type option for semester, queue and MOC
descriptions and for the call notes.  This is to allow
an administrator to select a formatted type as an alternative
to plain text in these locations.  RST has been added as
an format type, using the existing formatting function
which is used in the help system.

The FormatType class now keeps track of which formats are
available to users (current not used anywhere, but potentially
in proposal text and institution / person profiles) and which
are reserved for system usage.  System usage in this case
refers to the description elements mentioned above, as these
are maintained by administrators and form part of the web
interface.  RST is currently reserved for system usage.
(It may require more security before being made available to users.)