db: Record authentication fails to limit attempts
This adds a table "auth_failure" which is used to record the number
of failed authentication attempts for a given user name.  An expiry
time is set on the time of the first failure.  If an excessive number
of failures happen before this time passes, further attempts to
authenticate are rejected.