jcmt: Disable "array" ACSIS modes for non-array Rx
The JavaScript code now attempts to prevent the selection of an
"array-only" mode when a non-array receiver is chosen.  The
array-only mode options (in the selection control) are disabled
and if one was selected, the previous option is chosen instead.
(It is assumed that array-only options are always preceded by
the equivalent non-array-only option.)

Note that these modes aren't really "array-only".  They are the
2-subsystem hybrids which necessitate a lower frequency resolution
when used with an array receiver (namely HARP) because there is
no spare capacity in ACSIS for chaining.