web: Record referrer when logging in to take admin
When taking administrative privileges, with the new session expiry
system (or if the user logged out in another tab) it's possible that
there will be a redirect via the "log in" page.  This would make
the "referrer" for the "take admin" page be the "log in" page, but
going there is an error when the user is already logged in.  Therefore
add an extra option "record_referrer" to the "require_auth" decorator
to have it record the referrer in the session as "log_in_referrer"
if log in is required and clear it otherwise.  Then after taking
administrive privileges, we redirect to "log_in_referrer" in preference
to the actual referrer if it exists.

Note: because the variables controlling the log in process are stored
in the session, it's still possible to end up in the wrong place
(including the "log in" page when logged in) if you open multiple
"log in" pages at once.