jcmt: Add links to calculator documentation
Added a new section to the main "layout.html" template so that
if "help_link" is defined, then a link is displayed at the bottom
of the page.  If "help_link_text" is also defined then it is used
as the text for the link, otherwise "Help on this page" is used.