web: Move proposal submit link "beside" info table
The "Submit proposal" link (and "Withdraw proposal" / "Validate proposal"
links) have been moved to a borderless table cell beside the
proposal status row of the proposal information table.  This is
to give them a more prominent placement, different from that of the
other edit links.

The links are also decorated with an arrow, cross and tick mark
respectively.