web: Add note on proposal view for unprocessed PDF
Now if the proposal view page shows that a PDF file has not yet
been processed, there is also a note indicating that the user
should reload the page later.  This is to make it clear that
the fact that the PDF file hasn't yet been processed isn't an
error.