web: Add dashboard callout box for new proposals
When creating a new proposal or accepting a new invitation,
add a (hopefully unmissable) "callout" box indicating the
dashboard in the menu as the way to return to the proposal later.

Since the callout box is based in the "navbar" menu, the menu
can no longer use "overflow: auto" as the callout box should
extend below it.  Therefore use an :after pseudo-element to
clear the floats in the menu rather than "overflow: auto".