db/web: Add "department" to institution records
While a list of institutions with which to pre-fill the instition
table was being prepared, it was found that many of them did not
fit the permitted name / organization pattern.  Therefore
added a third level: organization.  This should allow us to more
usefully abbreviate institutions where necessary.  However for
now the institution should be shown in full everywhere.