web: Use timer to reduce JavaScript word counts
Now when the input event is fired, start a timer rather than
performing a word count immediately.  Subsequent inputs are ignored
while the timer is already running.  The timer is 500ms for short text
(<= 250 words) and 2s for longer text.  The idea is that the word count
(which can be slow on machine where JavaScript runs less quickly) should
not introduce excess delays when text is being typed in, so this
system prevents word counts being performed more frequently than the
defined time intervals.  The count isn't just done at a (repeating)
"interval" because then it would be re-computed even when nothing is
happening to the text.