jcmt: Change representative tau in Band 5 to 0.25
The representative 225 GHz opacity values used by the old
SCUBA-2 and Heterodyne (HiTEC) ITCs were inconsistent in
Bands 1 and 5.

SCUBA-2 (and Hedwig initial values):
    Band 1: 0.045
    Band 5: 0.23
HiTEC:
    Band 1: 0.04
    Band 5: 0.25

This is a problem for Hedwig as we only have one set of representative
opacity values for both ITCs.

JGW recommended 0.045 for Band 1 and 0.25 for Band 5, therefore this
commit changes the value for Band 5 to 0.25.