web: Convert calculator proposal_id arg to int
The proposal_id should be an integer, so convert it immediately.  This
allows us to compare it with other proposal identifiers when trying to
find the proposal code.  Probably not noticed before since the proposal
identifier would go through another HTTP request (a POST) before being
used to store the calculation.