db: Add table to record review committee decisions
This table is primarily to provide the "accept" boolean field,
indicating whether the committee has decided to accept the proposal.
This is separate from the ACCEPTED / REJECTED proposal states in
order to allow the committee to record decisions as they are being
made and to change them as necessary.  Only when the review is
complete (and the decision marked "ready" in this table) should
the proposal state be changed to its final state and the members
of the proposal informed of the outcome.