web: Special handling of auth. for POST requests
When a user is required to authenticate as the result of a POST
request, now show a customized log-in page with instructions
for how they should proceed.  They can either log in in a new
tab or log in and then navigate back to their form.

In both cases the "log_in_for" session variable is set to the
URL of a new page which thanks the user for logging in again
and instructs them to navigate back to their form.

The new behavior isn't ideal, but the problem is that at the point
of receiving a POST request for which authentication is not present,
it's too late for anything graceful to be easily done about it.
The new page should definately be an improvement over not trapping
POST here at all (as was previously the case) because then the user
would, on logging in, be redirected back to a copy of the form they
were editing but without all their edits.