web: Label private MOCs in the clash tool
Now matches and MOC list entries have a "private" label if the
coverage map is not public.  There is also a "Visibility: Private"
entry in the additional information table on the coverage map
view page.

This is to make it clear to technical assessors (and administrators)
that a proposal author would not have been able to see these clash
tool results.