web: Add more information to tabulation CSV file
A method to get the titles and a generator method for rows of the
CSV file have been added so that they can be overriden by facility-
specific classes.  The main "view_review_call_tabulation_download"
method can append the non-PI members to the end of each row --- this
is optional but currently always enabled.