web: Use white background for all tables
Previously this was only specified for tables inside "div.hovertable".

This is intended to help when a table is too wide for the content "div"
so that it spills out over the silver background.  Now the table cells
will have a white background so that they are more readable in this
situation.