jcmt: Add instrument to JCMT time allocation
Now the JCMT time allocations include the instrument (as well
as the weather band).  This means that we can now handle
JCMT time allocations in the same way as observing requests
since they have the same structure.  Therefore now re-use the
JCMTRequest classes for allocations.  Also extracted the
request editing table into a template macro so that it can
be placed on the decision page for JCMT.

Requested by IMC (change to previous request that we only input the
allocated time by weather band).