jcmt: Skip blank times in time allocation
This is so that, when no time is being allocated, clicking the
"Save" button without deleting the default empty row in the alocation
table isn't an error.  For editing observing requests on proposals
(the other place this code is used) we do want to raise an error
if any of the time requests is left empty.