jcmt: Show allocation in tabulation when defined
Now whenever a decision is defined (accept or reject) the online
tabulation view shows the committee-allocated times rather than
the original request.  (The behavior now matches that described
at the top of the page.)  This means that the "total as shown"
row now gives the total of the allocations, where defined, even
for projects which are currently marked as being rejected.

Note the allocations are still hidden in the CSV file version of the
tabulation when the proposal is not marked as being accepted.