jcmt: Default time allocation to observing request
Now pre-fill the time allocation form with the times from the original
observing request.  This requires writing all of the requests with no
identifier (request.id=None) so that they appear in the form as, e.g.,
"time_new_1".

This required a new create_counter template function because a counter
implemented with {% set %} directives does not persist outside of its
loop and we need to pass the next value to the JavaScript via a data
attribute.