web: Check new pass. not blank before using token
We need to ensure that the password is valid before the reset token
is used because using the token removes it from the database
whereas we want the user to be able to try again with an acceptable
password.

Also added a note about this to the docstring for the
update_user_password database method.