web: Add ability to sort columns of tabulation
Now each row of the tabulation contains a data attribute with a
JSON object containing all of the values by which the table can
be sorted.  Then the corresponding table header cells are given
the "sortable" class and a data attribute indicating which value
is to be used to sort the table.  The sorting behavior is
fine-tuned via the "sortreverse" and "sortedalready" classes.

The JavaScript code used to sort the tableis based on that used by Crab,
with added support for null values -- these appear last whichever
order a column is being sorted in.