test: Add test reproducing sync uniqueness problem
The general _sync_records method doesn't currently consider
database uniqueness constraints.  This can lead to an error when
we want to sync an update which re-uses unique values from other
rows which we intend to alter / delete at the same time.  This
additional test method causes this error to occur.