db: Defer rows with uniq. problem in _sync_records
Now for records where applying an update would cause a problem with
a unique constraint, the update is put back to the end of the updates
queue.  We only need to check for collisions in the collection of
not-yet-updated existing values since we already check that the input
values do not violate the uniqueness constraint.

In the event of circular updates (e.g. A->B, B->C, C->A), the values
would be returned to the queue multiple times.  This situation is
detected and, for now, an exception is raised.  An additional unit
test has been added to test this behavior.