db: Improve update loop detection in _sync_records
Sometimes an update may need to be deferred more than once without
there actually being a complete circular update loop.  For example:

    (1) A -> B
    (2) B -> C
    (3) C -> D

On the first attempt we can't apply (1) or (2).  Then on the second
attempt, (2) is applied, since (3) removed the conflict with "C".
We can finally apply (1) on the third attempt.

Now the circular update situation is detected only if we attempt to
return to the same element without any updates having succeeded in
the meantime.  An example of this situation is:

    (1) A -> B
    (2) B -> C
    (3) C -> A