type: Mark EPS format not allowed for user figures
This is to (perhaps temporarily) disallow uploading figures in
EPS format.  Support for EPS format was originally requested by
an external tester but could cause problems for generation of
proposals in PDF format.  So it is being disabled until the method
of generating PDF files for proposals has been established.