web: Avoid escaping PDF URL page=N fragment
When generating the URL for a page of a PDF file, now add the
page=N fragment manually rather than specifying it via the Flask
url_for _anchor argument.  This is to prevent the = symbol from
being %-encoded since Firefox's PDF viewer seems not to recognise
the page specification when this happens.