web: Allow special edit access to feedback review
Added a special case to the hedwig.view.auth.for_review function
to allow all suitable reviewer roles to edit an existing feedback
review.  Note that this only applies when we check authorization
for a particular review and some templates don't know about the
special case so they don't offer links to edit the feedback review.