db: Update for separate insertion of MOC cells
This commit updates the database access code to separate the
writing and updating of MOC file metadata from the insertion of
the MOC cells into the "moc_cell" table.  This is because the
latter is a very time consuming step so it should be performed
offline in a poll process.

A new "state" column has been added to the "moc" table to allow
the processing status to be tracked.

MOC FITS file reading and writing routines have been extracted
into a new module, "hedwig.file.moc".

The "update_moc_cell" method (which is no longer private) now attempts
to perform an incremental update of the MOC cells rather than deleting
everything and re-inserting.  This should be more efficient in
the common case of wanting to update a large MOC to add more
observations.  It also performs the update as a series of separate
transactions so that the database lock is released between them.
Additional delays should be added here to allow other database
access to proceed while a MOC is being updated.

The methods "add_moc" and "update_moc" now expect a MOC object
rather than a buffer containing a MOC in FITS format (as they
did before commit 725237e922711e5b06dc74a569cf2f0f737bd386).
They they write the MOC back to FITS format before storing it
in the "moc_fits" table.  The reason for this is to make it so
that the normalized MOC (at the facility-specific desired level)
is written to the database.  At the poll stage we will then
be able simply to read the MOC back from the "moc_fits" table
and insert its cells into the database, without having to worry
about the MOC level.  It also ensures that the MOCs which users
may download will match exactly what the clash tool is using.