web: Add a page showing feedback for a proposal
This page is intended to allow the proposal members to see the
feedback review and time allocation online.  (This should be the
same information as is sent by email once the review committee
decision is ready.)

A link is shown to the feedback next to the status on the proposal
view when appropriate.  The link goes to a page showing the feedback
comments and, in the case of JCMT, the time allocation.