web: Apply with_verified_admin in people view
Now that the people view is a class rather than a function
package, it can make use of our method decorators.  This commit
applies the "with_verified_admin" decorator and also updates the
"require_auth" decorator so that it can be used to eliminate
the authorization code in the "take_admin" method.