web: Tidy proposal/review list view methods
Now these are separated into two "public" methods each, which call
a private method to prepare the template context.  This allows us
to apply method decorators to streamline the methods.

This commit includes adding a permission="none" option to the
"with_person" decorator, which works like the similar option for
the "with_proposal" decorator.