db: Add computed state to call search results
This adds a new enum-style type, "CallState", which is used to represent
the possible states a call can be in: unopened, open or closed.  The new
value is defined by comparing the current date to the call's opening
and closing dates.  The web interface has been updated to make use
of the new search result.

The search_call method now accepts a state argument rather than an
"is_open" argument, in order to allow for more flexibility.  When this
argument is defined, the method uses the SQLAlchemy "column" function
to generate a reference to the computed column.  (There may be a better
way to do this, but defining the computed column as a Python variable
and then adding it to both the select and where clause causes the case
expression to be duplicated in the SQL statement generated.)