web: Add a page listing closed calls for proposals
A link to this page appears on the facility home page when there
are semesters with closed calls but no open calls.  The closed calls
link to the semester pages, which currently give an error, but the
intention is to allow the call text for previous semesters to be
visible.