admin: Abort closing proposal in unexpected state
If a proposal is in an unexpected state, skip processing it entirely
when closing the call.  (Rather than only skipping changing its state
but still freezing its members' institutions.)  This is so that we don't
end up repeatedly freezing institutions for a proposal in the event of
a problem with the state information.