admin: Add poll function to close completed calls
This adds a new poll option "close" to the hedwigctl script.  It searches
for calls which have proposals in open states.  These states are identified
by a new method ProposalState.open_states, which currently assumes that
editable states are open.  If this ever changes (e.g. if submitted proposals
became non-editable) then the open_states method will have be adjusted.
A grace period has been added to the configuration file -- a call will be
closed this many minutes after the specified closing date.