db: Add note to decision table & allow null accept
The JCMT TAC requested the ability to record a note alongside each
decision (for communication with the observatory scheduler).

A note column has been added to the decision table.  In order to allow
these notes to be saved if the decision status is returned to undecided,
we must allow the decision.accept column to be null rather than deleting
the whole entry in the decision table, as was done before.