db: Don't refer to computed state in where clause
The where clause shouldn't refer to aliases, so we can't refer to
the computed "state" value for calls and messages.  (While SQLite
allowed this, MySQL does not.)

Therefore adjust the state handling recently added for calls
and messages (commits 8b82175e6bcc436f49711e8a1a77f87d39bcf956
and 91287ebf2d3235f98aecbfde6360887257a547b0) so that the
SQLAlchemy expression for the state is used in the select
statement and also directly re-used in the where clause.