web: Add custom error handler for common errors
This adds an error handler for "Not Found", "Forbidden" and
"Internal Server Error" exceptions to the Flask application.
This handler shows the error on a Hedwig-styled page (using
the existing "error.html" template.

This change should make certain errors, especially "Forbidden"
prages, more convenient to deal with as the Hedwig menu will
be shown, including the log in and take/drop admin controls.