web: Add mode to calculator convert_input_version
This method now takes the calculator mode as its first argument.
I think that when this method was first set up, it was assumed
that it could infer the mode from the inputs.  However since the
mode is readily available, it might be more convenient to give
it as a method argument.