type: Add SectionedList type
This class is intended to behave like a list but to store its
items in a number of sections.  This is to allow operations such
as:

* Iterating over the list, section-by-section.

* Adding items to a section while retaining the overall ordering
  of the sections.

Sections have a (typically short string) identifier and also
(optionally) a "name" to be used for display purposes.
This prevents the long name having to be used in each
method call and allows the names to be changed without
the section identifiers needing to be altered.