web: Explicitly track review edit referrer
Now use explicitly-set short referrer codes (tab = tabulation,
cr = call reviews, pr = proposal reviews) to identify pages which
link to the review edit page.  These are used to generate the
redirection target on successful saving of the review, in
place of the actual referrer URL from the HTTP header.

This is to avoid problems where the log in page appears when
the user navigates to the review edit page.  In this case the
log in page would be the referrer and redirecting back there
causes an "already logged in" error page to be displayed.