web: Handle errors other than HTTPExceptions
Flask sometimes calls our error handler with exceptions which aren't
HTTPException instances, e.g. when handling an uncaught exception.
This only happens when not in debug mode, as otherwise the debugger is
invoked.  Therefore before using the exception's attributes, check that
those which we need are present.  If not, display a standard HTTPError
(internal server error) response instead.