web: Add option to send reminder to ext. reviewer
Requested by the JCMT TAC.

This adds a new type of message which is the reminder, as opposed
to the existing re-invite message, the purpose of which is to send
a new invitation code.  A reminder could perhaps just refer to the
previous message, but it was thought that it might be useful to
include all the information from the original invitation, including
the invitation code (if it is still valid).  Therefore the reminder
email uses the same template as the invitation email with a context
parameter to distinguish them.  The confirm pages for the two kinds
of message have been written to try to clarify the difference
between them.

This commit includes a draft of the introductory paragraph for the
reminder email, but this text should be approved before use is
made of it.