db: Add message "discard" flag
This adds a new boolean flag to the message table which can be used
to mark a message as discarded.  When this is set, the state of the
message should be returned as DISCARD and the message should not
be retrieved by the "get_unsent_message" method.  The intention is
to use this feature to mark messages where sending failed but
where the message is no longer relevant, so we do not want to
try sending it again.