db: Add message thread information
This commit adds thread "type" and "id" columns to the message table.
The type indicates why the messages should be grouped together, while
the "id" identifies the particular thread.  It will usually match
the identifier of some other database object.  For example threads
about proposals would use the proposal identifier here.  When a message
is selected to be sent, the resulting tuple now contains a list of
previous message identifiers (Message-ID headers) for the same thread.