type: Add new "final review" state
This adds a new proposal state between the "review" and the final
"accepted" and "rejected" states during which decisions can be
entered.

The JCMT TAC have requested that TAC member ratings be hidden
until their meeting, which, since access to proposals and reviews
is controlled by proposal state, necessitates a new state to be
used during the meeting.