db: Combine actions in use_password_reset_token
This database method now includes actions which were previously
performed in the web view method: changing the password, looking
up the user name and clearing the authentication failure log.
We now only log the usage of the token, and don't separately log
the password being changed.  The reason for this change is to reduce
the risk of an error between the token being used and the password
being changed leading to the token being deleted prematurely.