db: Use separate member field in search_proposal
Now when search_proposal is given a "person_id" or called with the
"person_pi" option, the corresponding member information object is
stored in the Proposal tuple's (new) "member" field rather than the
"members" field.  This is to avoid the previous confusing behavior
where the "members" field would sometimes contain only a single
information object (rather than a member collection) depening on
the arguments to "search_proposal".