db: Use separate reviewer field in search_proposal
When search_proposal is called with the reviewer_person_id argument,
it now returns the corresponding reviewer information in the
proposal.reviewer field rather than sharing the proposal.reviewers
field with reviewer collections.  This is similar to the recent
change of returning a single proposal member in the reviewer.member
field to avoid confusion.