web: Show affected people on inst. edit confirm
In order to discourage people from editing their current institution
when they move, display an actual list of the people who will be
affected by the edit.  (I.e the current members of the institution.)
This is shown as a list of those with public profiles plus the
number of private / unregistered people.