web: Skip rev. auth for person unless rev. coord.
If the user is not a review coordinator for any queues, when checking
authorization for an unregistered person profile, there's no need to
search for reviewer records for that person.

When that was done, "search_reviewer" was called with an empty list
of queues and SQLAlchemy warned that an in predicate on "call.queue_id"
was invoked with an empty sequence.  Instead we should just skip this
part of the authorization function when the user isn't a review
coordinator.