test: Add option to sim. state update in auth test
This new option (an instance attribute), if set, allows some of
the tests to simulate proposal state updates by modifying the
state in the proposal tuple which is going to be used to call
auth module functions.  This gives improved performance over
actually updating the state in the database and the querying
again to retrieve the updated proposal tuple.