web: More facility-specific overall rating calc.
Now the ReviewerCollection.get_overall_rating method takes a function
to be called to obtain the rating and wieght to be used in the calculation
for each completed review.  The generic facility view class supplies
a function which retains the previous behavior, treating unweighted
reviews as having a weighting of 100%.  The JCMT facility view class
now uses the JCMT reviewer expertise level to determine the weighting.