web: Re-inline tabulation page review row lines
Now render directly from the template rather than via a macro.
This is done by using a loop over values for the "cttee_role"
flag, rather than calling the macro for each value.  The reason
for this change is to make it easier to customize the review rows
for specific facilities.