jcmt: Add ancillary instrument to req. / alloc.
This adds a new enum-style class for JCMT ancillary instruments
which corresponds to the new "ancillary" field in the jcmt_request
and jcmt_allocation database tables.  (This makes the assumption
that at most one ancillary instrument would be in use at any time.)
A value of 0 is used to indicate that no ancillary instrument is
being used, rather than None, because None would be represented in
the database by NULL, which would prevent the uniqueness constraint
on instrument, ancillary and weather band being applied correctly.

The JCMTInstrument class now has a list of ancillary instruments
which can be used with each main instrument.  (It is assumed that
the instrument can always also be used without any ancillary at all.)

The request and allocation editing pages and the to_sorted_list
method used to prepare information for the feedback message have
been updated to take account of ancillary instruments.