util: Add function to identify list-like types
This checks if the value is a list or tuple.  This saves repeating
the same test multiple times and will make it easier to allow for
more iterable types.  The reason for checking against known types
(list and tuple) is to be sure to exclude strings, which might
appear iterable.