db: Add statement sequence generator method
This new method is intended to make it easy to write query routines which
select results via an "in" constraint without worrying about how long the
specified list of values is.  A sequence of statements is returned after
breaking the list into blocks and applying an "in" constraint for each
block.  As an additional convenience, if no constraint is being specified,
the query is yielded as is -- this is to avoid the need for different code
in the cases of there being and not being an "in" constraint to apply.