jcmt: Gather extra tabulation information together
Now when gathering information for the detailed tabulation review
page, make a list of all of the proposal_id values appearing on the
tabulation.  Then search for the requests, allocations and options for
all the proposal_id values together.  This should reduce the number of
JCMT-specific queries required to prepare the tabulation page.  It
relies on the new "CollectionByProposal" mix-in to allow information
to be extracted from the results collections on a proposal-by-proposal
basis.